mac80211: backport of r35671: b43: Increase number of RX DMA slots
Bastian Bittorf reported that some of the silent freezes on a Linksys WRT54G
were due to overflow of the RX DMA ring buffer, which was created with 64
slots. That finding reminded me that I was seeing similar crashed on a netbook,
which also has a relatively slow processor. After increasing the number of
slots to 128, runs on the netbook that previously failed now worked; however,
I found that 109 slots had been used in one test. For that reason, the number
of slots is being increased to 256.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>


git-svn-id: svn://svn.openwrt.org/openwrt/branches/attitude_adjustment@35947 3c298f89-4303-0410-b956-a3cf2f4a3e73